+++ updater/software/All.cs
@@ -59,7 +59,8 @@ namespace updater.software
                 new OpenJRE8(autoGetNewer),
                 new OpenJRE11(autoGetNewer),
                 new OpenJRE17(autoGetNewer),
-                new OpenJRE21(autoGetNewer)
+                new OpenJRE21(autoGetNewer),
+                new OpenJRE25(autoGetNewer)
             };
 
             // Firefox (release channel)